:bug: Fix deepscan issues
